Extract resolvers and mutations for discount
